--- packages/mui-material/src/styles/md3/defaultTheme.ts
@@ -1,5 +1,0 @@
-import extendTheme from '../extendTheme';
-
-const defaultTheme = extendTheme();
-
-export default defaultTheme;
