--- packages/mui-material/src/styles/md3/index.ts
@@ -1,11 +1,0 @@
-'use client';
-export { rootShouldForwardProp } from '@mui/material/styles/styled';
-export * from './Theme.types';
-// export * from './extendTheme';
-// export { default as extendTheme } from './extendTheme';
-export { default as styled } from './styled';
-export { default as defaultTheme } from './defaultTheme';
-export { default as useTheme } from './useTheme';
-export { default as shouldSkipGeneratingVar } from './shouldSkipGeneratingVar';
-export * from './CssVarsProvider';
-export { default as useThemeProps } from './useThemeProps';
